refactor: improve module structure
