+++ .idea/gradle.xml
@@ -9,6 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/core" />
+            <option value="$PROJECT_DIR$/example-bot" />
           </set>
         </option>
       </GradleProjectSettings>
