+++ core/src/main/kotlin/data/MessagePayload.kt
@@ -1,0 +1,9 @@
+package me.hechfx.foreground.core.data
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class MessagePayload(
+    val uri: String,
+    val cid: String
+)
