+++ core/src/main/kotlin/entity/Entity.kt
@@ -1,0 +1,4 @@
+package me.hechfx.foreground.core.entity
+
+interface Entity {
+}
