+++ core/src/main/kotlin/entity/vanilla/user/User.kt
@@ -1,0 +1,8 @@
+package me.hechfx.foreground.core.entity.vanilla.user
+
+import me.hechfx.foreground.core.entity.*
+
+class User(
+    val did: String,
+    val asMention: String
+) : Entity
