+++ core/src/main/kotlin/events/ForegroundEvent.kt
@@ -1,0 +1,5 @@
+package me.hechfx.foreground.core.events
+
+abstract class ForegroundEvent {
+    abstract val INTERNAL_NAME: String
+}
