+++ core/src/main/kotlin/events/vanilla/identity/IdentityEvent.kt
@@ -1,0 +1,12 @@
+package me.hechfx.foreground.core.events.vanilla.identity
+
+import com.upokecenter.cbor.CBORObject
+import me.hechfx.foreground.core.events.*
+
+class IdentityEvent(raw: CBORObject) : ForegroundEvent() {
+    override val INTERNAL_NAME = "#identity"
+
+    val did = raw["did"].AsString()
+    val seq = raw["seq"].AsInt32()
+    val handle = raw["handle"].AsString()
+}
