+++ core/src/main/kotlin/utils/BSConstants.kt
@@ -1,0 +1,6 @@
+package me.hechfx.foreground.core.utils
+
+object BSConstants {
+    const val BASE_URL = "https://bsky.app"
+    const val BASE_API_URL = "https://bsky.social/xrpc"
+}
