+++ example-bot/src/main/kotlin/ExampleBot.kt
@@ -1,0 +1,22 @@
+package me.hechfx.bot
+
+import me.hechfx.bot.listeners.*
+import me.hechfx.bot.utils.config.*
+import me.hechfx.foreground.core.gateway.ForegroundWebSocketClient
+
+class ExampleBot(
+    val config: ExampleBotConfig
+) {
+    lateinit var client: ForegroundWebSocketClient
+
+    suspend fun start() {
+        client = ForegroundWebSocketClient(
+            config.identifier,
+            config.password
+        )
+
+        MajorListener(this).listen()
+
+        client.connect()
+    }
+}
