+++ example-bot/src/main/resources/config.conf
@@ -1,0 +1,2 @@
+identifier="your-user.bsky.social"
+password="your-password"
