+++ settings.gradle.kts
@@ -1,5 +1,18 @@
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
 }
+
 rootProject.name = "foreground"
+
+dependencyResolutionManagement {
+    versionCatalogs {
+        create("libs") {
+            from(files("foreground.versions.toml"))
+        }
+    }
+}
+
+
+include("core")
+include("example-bot")
