--- src/main/kotlin/me/hechfx/foreground/data/MessagePayload.kt
@@ -1,9 +1,0 @@
-package me.hechfx.foreground.data
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class MessagePayload(
-    val uri: String,
-    val cid: String
-)
