--- src/main/kotlin/me/hechfx/foreground/entity/Entity.kt
@@ -1,4 +1,0 @@
-package me.hechfx.foreground.entity
-
-interface Entity {
-}
