--- src/main/kotlin/me/hechfx/foreground/entity/vanilla/user/User.kt
@@ -1,8 +1,0 @@
-package me.hechfx.foreground.entity.vanilla.user
-
-import me.hechfx.foreground.entity.*
-
-class User(
-    val did: String,
-    val asMention: String
-) : Entity
