--- src/main/kotlin/me/hechfx/foreground/events/ForegroundEvent.kt
@@ -1,5 +1,0 @@
-package me.hechfx.foreground.events
-
-abstract class ForegroundEvent {
-    abstract val INTERNAL_NAME: String
-}
