--- src/main/kotlin/me/hechfx/foreground/events/ForegroundEventType.kt
@@ -1,19 +1,0 @@
-package me.hechfx.foreground.events
-
-enum class ForegroundEventType(val raw: String) {
-    COMMIT("#commit"),
-    ACCOUNT("#account"),
-    IDENTITY("#identity"),
-    HANDLE("#handle"),
-    TOMBSTONE("#tombstone");
-
-    companion object {
-        fun from(raw: String): ForegroundEventType {
-            return try {
-                entries.first { it.raw == raw }
-            } catch (e: NoSuchElementException) {
-                throw NoSuchElementException("ForegroundEventType not found for '$raw'")
-            }
-        }
-    }
-}
