--- src/main/kotlin/me/hechfx/foreground/events/vanilla/account/AccountEvent.kt
@@ -1,14 +1,0 @@
-package me.hechfx.foreground.events.vanilla.account
-
-import com.upokecenter.cbor.CBORObject
-import me.hechfx.foreground.events.*
-import java.time.*
-
-class AccountEvent(raw: CBORObject) : ForegroundEvent() {
-    override val INTERNAL_NAME = "#account"
-
-    val did = raw["did"].AsString()
-    val seq = raw["seq"].AsInt32()
-    val time = Instant.parse(raw["time"].AsString())
-    val active = raw["active"].AsBoolean()
-}
