--- src/main/kotlin/me/hechfx/foreground/utils/BSConstants.kt
@@ -1,6 +1,0 @@
-package me.hechfx.foreground.utils
-
-object BSConstants {
-    const val BASE_URL = "https://bsky.app"
-    const val BASE_API_URL = "https://bsky.social/xrpc"
-}
